+++ packages/frontend/lib/auth/login-user.ts
@@ -1,0 +1,3 @@
+import { LoginUserFormData } from "../types/auth/user-form";
+
+export async function loginUser(data: LoginUserFormData) {}
